Refactor baseband sample processing functions into own file
