+++ include/rtl_433.h
@@ -31,7 +31,6 @@
 #define DEFAULT_DECIMATION_LEVEL 0
 #define MINIMAL_BUF_LENGTH      512
 #define MAXIMAL_BUF_LENGTH      (256 * 16384)
-#define FILTER_ORDER            1
 #define MAX_PROTOCOLS           25
 #define SIGNAL_GRABBER_BUFFER   (12 * DEFAULT_BUF_LENGTH)
