+++ src/CMakeLists.txt
@@ -18,6 +18,7 @@
 ########################################################################
 add_executable(rtl_433 
 	rtl_433.c
+    baseband.c
 	bitbuffer.c
 	pulse_demod.c
 	pulse_detect.c
